Volume integral and test implementation.
Some errors were fixed while trying to got the induction prediction working.
The volume integral was correctly implemented when trying to run the code for the base layer.
The test script was added to check if the code is working as expected. It goes through the same
pipeline as the main script.